fix: Simpler colour schemes
If you don't choose a colour scheme, the site switches
based on device preferences.
A toggle button in the header allows users to override the chosen
scheme.

If you choose a light or dark colour scheme in the config file,
the toggle is hidden and that scheme is forced throughout.